added int()

+++ supervised_learning/0x07-cnn/2-conv_backward.py
@@ -16,8 +16,8 @@ def conv_backward(dZ, A_prev, W, b, padding="same", stride=(1, 1)):
     tmp = A_prev
 
     if padding is 'same':
-        ph = (((sh * h_prev) - sh + kh - h_prev) // 2)
-        pw = (((sw * w_prev) - sw + kw - w_prev) // 2)
+        ph = int((((sh * h_prev) - sh + kh - h_prev) // 2))
+        pw = int((((sw * w_prev) - sw + kw - w_prev) // 2))
     if padding is 'valid':
         ph, pw = 0, 0
